+++ examples/react-router-7/app/routes/_index.tsx
@@ -1,0 +1,5 @@
+import { LiFiWidget } from '../components/LiFiWidget'
+
+export default function Index() {
+  return <LiFiWidget />
+}