+++ examples/reown/src/providers/WalletProvider.tsx
@@ -1,18 +1,22 @@
-import { useSyncWagmiConfig } from '@lifi/wallet-management'
 import { ChainType, type ExtendedChain, useAvailableChains } from '@lifi/widget'
-import { bitcoin, solana } from '@reown/appkit/networks'
-import { type AppKit, createAppKit } from '@reown/appkit/react'
+import { bitcoin, mainnet, solana } from '@reown/appkit/networks'
+import {
+  type AppKit,
+  createAppKit,
+  useAppKitAccount,
+} from '@reown/appkit/react'
 import { BitcoinAdapter } from '@reown/appkit-adapter-bitcoin'
 import { SolanaAdapter } from '@reown/appkit-adapter-solana'
 import { WagmiAdapter } from '@reown/appkit-adapter-wagmi'
 import type { AppKitNetwork } from '@reown/appkit-common'
+
 import { useEffect, useRef } from 'react'
 import { WagmiProvider } from 'wagmi'
 import { metadata, projectId } from '../config/appkit'
 import { chainToAppKitNetworks, getChainImagesConfig } from '../utils/appkit'
-import { emitter, SolanaProvider } from './SolanaProvider'
+import { SolanaProvider } from './SolanaProvider'
 
-export function WalletProvider({
+export function ReownEVMWalletProvider({
   children,
   chains,
 }: {
@@ -51,49 +55,38 @@ export function WalletProvider({
       metadata,
       chainImages,
       themeMode: 'light',
+      defaultNetwork: mainnet,
     })
-
     wagmi.current = wagmiAdapter
     modal.current = appKit
   }
 
   const { wagmiConfig } = wagmi.current
 
-  useSyncWagmiConfig(wagmiConfig, [], chains)
+  const { isConnected, status } = useAppKitAccount()
+  const { isConnected: evmIsConnected } = useAppKitAccount({
+    namespace: 'eip155',
+  })
 
+  const { setCaipNetwork, getCaipNetwork } = modal.current
+
+  // In multichain mode, Appkit fails to update the main connection state correctly after one chain is disconnected
+  // If there is mismatch between main connection state and eth connection state, we manually update the main state
+  // Details here: https://github.com/reown-com/appkit/issues/5066
   useEffect(() => {
-    const appKit = modal.current
-    if (appKit) {
-      const unsubscribe = appKit.subscribeNetwork((network) => {
-        if (network.caipNetwork) {
-          const { chainNamespace } = network.caipNetwork
-          if (chainNamespace === 'solana') {
-            const connectors = appKit.getConnectors(chainNamespace)
-            // We use the first connector in the list as there's no way to get the active connector from appKit yet.
-            emitter.emit('connect', connectors[0].name)
-          }
-        }
-      })
-
-      return () => {
-        emitter.emit('disconnect')
-        unsubscribe()
-      }
+    if (!isConnected && status === 'disconnected' && evmIsConnected) {
+      setCaipNetwork(getCaipNetwork('eip155'))
     }
-  }, [])
+  }, [isConnected, status, evmIsConnected, setCaipNetwork, getCaipNetwork])
 
   return (
-    <WagmiProvider config={wagmiConfig} reconnectOnMount={false}>
+    <WagmiProvider config={wagmiConfig} reconnectOnMount={true}>
       {children}
     </WagmiProvider>
   )
 }
 
-export function SyncedWalletProvider({
-  children,
-}: {
-  children: React.ReactNode
-}) {
+export function WalletProvider({ children }: { children: React.ReactNode }) {
   // fetch available chains before rendering the WalletProvider
   const { chains, isLoading } = useAvailableChains()
 
@@ -106,8 +99,8 @@ export function SyncedWalletProvider({
   }
 
   return (
-    <WalletProvider chains={chains}>
+    <ReownEVMWalletProvider chains={chains}>
       <SolanaProvider>{children}</SolanaProvider>
-    </WalletProvider>
+    </ReownEVMWalletProvider>
   )
 }