+++ packages/widget/src/hooks/useRouteExecution.ts
@@ -35,7 +35,7 @@ export const useRouteExecution = ({
   const queryClient = useQueryClient()
   const { account } = useAccount()
   const resumedAfterMount = useRef(false)
-  const { keyPrefix } = useWidgetConfig()
+  const { keyPrefix, sdkConfig } = useWidgetConfig()
   const emitter = useWidgetEvents()
   const routeExecutionStoreContext = useRouteExecutionStoreContext()
   const routeExecution = useRouteExecutionStore(
@@ -132,6 +132,7 @@ export const useRouteExecution = ({
         acceptExchangeRateUpdateHook,
         infiniteApproval: false,
         executeInBackground,
+        ...sdkConfig?.executionOptions,
       })
     },
     onMutate: () => {